Introduced CSVReceptor for writing data to a CSV file and ReceptorLambda for easily creating a Receptor from a lambda. Also various bug fixes and cleanup throughout. This is version 0.11.

+++ world/org/almostrealism/audio/AudioProteinCache.java
@@ -24,7 +24,7 @@ public class AudioProteinCache implements ProteinCache<Long> {
 	public static int sampleRate = 24 * 1024; // 44100
 	public static int bufferDuration = 100;
 	
-	public static int depth = Integer.MAX_VALUE;
+	public static long depth = Long.MAX_VALUE;
 	public static long convertToByte = depth / Byte.MAX_VALUE;
 	
 	private long data[] = new long[sampleRate * bufferDuration];
